reorganised tests to reflect which files funcitons are written in
