+++ tests/lephare/test_prepare.py
@@ -1,11 +1,10 @@
 import os
-import shutil
 
 import lephare as lp
 from lephare.prepare import all_types_to_keymap
 
 
-def test_prepare(test_data_dir):
+def test_prepare(test_data_dir: str):
     test_dir = os.path.abspath(os.path.dirname(__file__))
     os.environ["LEPHAREDIR"] = os.path.join(test_dir, "../data")
     os.environ["LEPHAREWORK"] = os.path.join(test_dir, "../tmp")
@@ -17,27 +16,6 @@ def test_prepare(test_data_dir):
     assert default_config["FILTER_REP"] == str(os.path.join(test_data_dir, "filt"))
     # Check it made the galaxy binary file
     assert os.path.exists(os.path.join(test_dir, "../tmp/lib_mag/CE_COSMOS.bin"))
-
-
-def test_load_sed_list(test_data_dir):
-    test_dir = os.path.abspath(os.path.dirname(__file__))
-    # Move one of the example sed folders
-    _ = shutil.copytree(
-        os.path.join(test_dir, "../data/sed/QSO"), os.path.join(test_dir, "../tmp/seds"), dirs_exist_ok=True
-    )
-    lp.load_sed_list(os.path.join(test_dir, "../tmp/seds/ONE_SED.list"), "QSO")
-    # Check the list is there
-    assert os.path.exists(os.path.join(test_dir, "../data/sed/QSO/ONE_SED/ONE_SED.list"))
-    # Check the sed is there
-    assert os.path.exists(os.path.join(test_dir, "../data/sed/QSO/ONE_SED/o5v.sed.ext"))
-
-    # Check it can run even if the file is already there
-    lp.load_sed_list(os.path.join(test_dir, "../tmp/seds/ONE_SED.list"), "QSO")
-    # Check absolute paths
-    lp.load_sed_list(os.path.join(test_dir, "../tmp/seds/ONE_SED.list"), "QSO", absolute_paths=True)
-    # Clear the copied folders
-    shutil.rmtree(os.path.join(test_dir, "../tmp/seds"))
-    shutil.rmtree(os.path.join(test_dir, "../data/sed/QSO/ONE_SED"))
 
 
 def test_all_types_to_keymap():
